pdu,dimse: fuzz the wire decoders; take 32-bit lengths out of int
pdu.Read runs on an open port before any AE title is checked, and everything
below it — the A-ASSOCIATE item walk, DecodePDataTF, and every dimse Decode* via
decodeElements — is reachable by anyone who can connect. All of it was covered
only by golden fixtures, which say what happens to well-formed input and nothing
about the rest.

Four targets: FuzzRead, FuzzDecodePDataTF, FuzzDecodeElements, and
FuzzCommandDecoders (the dispatch an SCP performs: Peek the Command Field, then
hand the same bytes to whichever decoder it names). Seeded from the pynetdicom
goldens plus the shapes they leave out — role selection and user identity, whose
length fields nest inside the item's; a multi-PDV PDU; an MCH-only fragment; an
N-ACTION command set.

They check more than absence of panics. Read must re-encode and re-read to the
same bytes, since the association code echoes negotiated items back to the peer.
DecodePDataTF must not produce an empty PDV value, which IsCommand, IsLast and
Fragment all silently misread. decodeElements accounts for every input byte, so a
length read differently from the way it was written stops adding up. Ran clean at
~86M executions across the four.

What the fuzzer could not reach on amd64: the lengths were read with
int(uint32), and int is 32 bits on 386/arm. A DIMSE element declaring Value
Length 0xFFFFFFF8 came out as -8, "off+8+vlen > len(b)" read as satisfied, and
the slice inverted — 8 bytes, no association state. A PDV item declaring
0x7FFFFFFF stayed positive but overflowed off+4+itemLen, same result from 20
bytes. Both now compare in uint64 before any conversion. pdu.go also could not
compile for a 32-bit target at all: errTooLong took its limit as an int and one
caller passes math.MaxUint32.

Neither panic was reachable in a released build, because godicom does not compile
for 32-bit either (encaps/encode.go:65). This does not make gonetdicom a 32-bit
library; it stops gonetdicom from being the reason.

Verified: GOARCH=386 now builds pdu and dimse and passes both packages' tests,
and each length test panics against the pre-fix guard there while passing on
amd64 either way. All four fuzz targets also ran clean for 40s each under
GOARCH=386 (-parallel 1; the engine cannot map its shared memory with more
workers in a 32-bit address space, and there is no coverage instrumentation for
that arch, so those runs are blind and short).

Closes part of #39 (fuzz targets)

+++ dimse/command.go
@@ -160,13 +160,16 @@ func decodeElements(b []byte) ([]element, error) {
 	for off+8 <= len(b) {
 		group := binary.LittleEndian.Uint16(b[off : off+2])
 		elem := binary.LittleEndian.Uint16(b[off+2 : off+4])
-		vlen := int(binary.LittleEndian.Uint32(b[off+4 : off+8]))
-		if off+8+vlen > len(b) {
+		// Value Length is 32 bits, so it is compared in uint64 rather than converted
+		// to int: where int is 32 bits wide a length above MaxInt32 becomes negative,
+		// the bounds check below passes, and the slice that follows panics.
+		vlen := uint64(binary.LittleEndian.Uint32(b[off+4 : off+8]))
+		if uint64(off)+8+vlen > uint64(len(b)) {
 			return nil, fmt.Errorf("dimse: truncated element (%04X,%04X)", group, elem)
 		}
-		val := append([]byte(nil), b[off+8:off+8+vlen]...)
+		val := append([]byte(nil), b[off+8:off+8+int(vlen)]...)
 		els = append(els, element{tag: uint32(group)<<16 | uint32(elem), value: val})
-		off += 8 + vlen
+		off += 8 + int(vlen)
 	}
 	if off != len(b) {
 		return nil, fmt.Errorf("dimse: trailing %d bytes", len(b)-off)
